[re-factory] split timeline file into components
