+++ src/components/Timeline/Guide.module.scss
@@ -1,0 +1,65 @@
+.btnContainer {
+	background-color: rgb(53, 52, 52);
+	color: white;
+
+	height: 36%;
+	width: 100%;
+
+	position: relative;
+	display: flex;
+	align-items: center;
+}
+
+@mixin btn-style {
+	height: 1.2rem;
+	width: 1.2rem;
+}
+
+.btnContainer__button {
+	@include btn-style;
+
+	fill: rgb(0, 174, 255);
+
+	cursor: pointer;
+	transition: all 0.2s;
+
+	&:hover {
+		transform: scale(1.1);
+	}
+}
+
+.btnContainer__button-disabled {
+	@include btn-style;
+	fill: rgb(95, 95, 95);
+}
+
+.btnContainer__left {
+	display: flex;
+	justify-content: space-around;
+	align-self: center;
+}
+
+.btnContainer__right {
+	position: absolute;
+	display: flex;
+	height: 100%;
+	flex-direction: column;
+	justify-content: space-evenly;
+	right: 1rem;
+}
+
+.btnContainer__icon {
+	height: 1.5rem;
+	width: 1.5rem;
+	margin-left: 1rem;
+}
+
+@media screen and (max-width: 1032px) {
+	.btnContainer__text {
+		display: none;
+	}
+}
+
+.btnContainer__text {
+	margin-left: 0.5rem;
+}
